+++ cli/docs/changelog.md
@@ -1,5 +1,11 @@
 # TBD
 ### Features
+* Added `service pause` and `service unpause` to CLI
+* Added E2E tests for pause and unpause for both Golang and Typescript
+
+### Changes
+* Upgraded `kurtosis core lib` to 1.45.4 which adds `EnclaveContext.PauseService` and `EnclaveContext.UnpauseService`
+* Upgraded `kurtosis engine lib` to 1.17.4
 * Added `uploadFiles` tests for typescript and go clients.
 
 ### Fixes